requirement.txt
beautifulsoup4==4.12.3
fastapi==0.103.2
pillow==11.0.0
requests==2.32.3
uvicorn==0.23.2